+++ src/app/dashboard/doctor/page.jsx
@@ -1,0 +1,3 @@
+export default function page() {
+    return <>This is doctor dashboard</>;
+}
