+++ src/app/dashboard/hospital/page.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const page = () => {
+    return <div>This is Hospital Dashboard</div>;
+};
+
+export default page;
