+++ src/app/page.js
@@ -3,7 +3,6 @@ import ProtectRoute from "@/components/ProtectRoute";
 export default function Home() {
     return (
         <>
-            <ProtectRoute />
             <h2>Wait a sec ...</h2>
         </>
     );
