+++ src/app/user/page.jsx
@@ -2,7 +2,7 @@ import Layout from "./Layout";
 export default function Page() {
     return (
         <Layout>
-            <h2>THis is user page</h2>
+            <h2>user page</h2>
         </Layout>
     );
 }
