+++ src/components/ProtectRoute.js
@@ -7,7 +7,9 @@ export default function ProtectRoute() {
     const user = JSON.stringify(Cookies.get("user_info"));
     console.log("token", token);
     if (!token) {
-        router.push("auth/register");
+        router.push("/auth/login");
+    } else if (token) {
+        router.push("/user");
     } else {
         if (user.role === "admin") {
             router.push("/dashboard");
